+++ changelog.md
@@ -2,6 +2,7 @@
 - Updated recent provider overlay
 - Fixed wallpaper app crashing issue after installing launcher for the first time
 - Updated apks & permissions
+- Added developer options
 
 **Version 1.2**
 - Added a workaround to fix material you issue happening in some roms